Move EventsIndicatorView and FoldableTextView into 'views' package
